Add files via upload
After completing my Google Data Analytics Certification, I started working on my first project which was about Bellabeat which is the go-to wellness brand for women with an ecosystem of products and services focused on women’s health. I looked at the business problem and try to come up with insights that will help the marketing team to come up with new marketing strategies.